msm-fb-refresher: Quit after 5 seconds by default when enabled
It appears that on my 4.4 downstream kernel OnePlus 5 the framebuffer
needs to be refreshed constantly or else the display will just blank
again; with this one can define an amount of time to keep the program
running until eventually reaching a distro userspace where e.g. the GPU
takes over.